Add LoggerTrait and refactor logging
