+++ Plugin/Catalog/Block/AbstractBlock.php
@@ -47,11 +47,15 @@ use Nosto\Cmp\Helper\Data as NostoCmpHelperData;
 use Nosto\Cmp\Helper\SearchEngine;
 use Nosto\Cmp\Model\Service\Recommendation\StateAwareCategoryService;
 use Nosto\Cmp\Model\Service\Recommendation\StateAwareCategoryServiceInterface;
+use Nosto\Cmp\Utils\Traits\LoggerTrait;
 use Nosto\Tagging\Helper\Account as NostoHelperAccount;
 use Nosto\Tagging\Logger\Logger;
 
 abstract class AbstractBlock extends Template
 {
+    use LoggerTrait {
+        LoggerTrait::__construct as loggerTraitConstruct; // @codingStandardsIgnoreLine
+    }
 
     /** @var SearchEngine */
     protected $searchEngineHelper;
@@ -70,9 +74,6 @@ abstract class AbstractBlock extends Template
 
     /** @var NostoHelperAccount */
     private $nostoHelperAccount;
-
-    /** @var Logger */
-    private $logger;
 
     /** @var string */
     public static $currentOrder;
@@ -101,11 +102,13 @@ abstract class AbstractBlock extends Template
         SearchEngine $searchEngineHelper,
         Logger $logger
     ) {
+        $this->loggerTraitConstruct(
+            $logger
+        );
         $this->categoryService = $categoryService;
         $this->paramResolver = $parameterResolver;
         $this->nostoCmpHelperData = $nostoCmpHelperData;
         $this->nostoHelperAccount = $nostoHelperAccount;
-        $this->logger = $logger;
         $this->storeManager = $context->getStoreManager();
         $this->searchEngineHelper = $searchEngineHelper;
         parent::__construct($context);
