+++ Plugin/Framework/Search/Request/GraphQlHandler.php
@@ -123,15 +123,7 @@ class GraphQlHandler extends AbstractHandler
     public function parseLimit(array $requestData)
     {
         if ($this->pageSize != -1) {
-            $this->getLogger()->debugWithSource(
-                sprintf(
-                    'Using DI value (%s) for the page size',
-                    $this->pageSize
-                ),
-                [],
-                $this
-            );
-
+            $this->debugWithSource(sprintf('Using DI value (%s) for the page size', $this->pageSize));
             return $this->pageSize;
         }
