+++ Plugin/Framework/Search/Request/RequestCleaner.php
@@ -39,6 +39,7 @@ namespace Nosto\Cmp\Plugin\Framework\Search\Request;
 use Exception;
 use Magento\Framework\Search\Request\Cleaner;
 use Nosto\Cmp\Utils\Search;
+use Nosto\Cmp\Utils\Traits\LoggerTrait;
 use Nosto\Tagging\Logger\Logger;
 
 class RequestCleaner
@@ -48,6 +49,10 @@ class RequestCleaner
     const KEY_CATEGORY_FILTER = 'category_filter';
     const KEY_QUERIES = 'queries';
     const KEY_FILTERS = 'filters';
+
+    use LoggerTrait {
+        LoggerTrait::__construct as loggerTraitConstruct; // @codingStandardsIgnoreLine
+    }
 
     /** @var GraphQlHandler */
     private $graphqlHandler;
@@ -70,9 +75,11 @@ class RequestCleaner
         GraphQlHandler $graphQlHandler,
         Logger $logger
     ) {
+        $this->loggerTraitConstruct(
+            $logger
+        );
         $this->webHandler = $webHandler;
         $this->graphqlHandler = $graphQlHandler;
-        $this->logger = $logger;
     }
 
     /**
@@ -87,10 +94,9 @@ class RequestCleaner
     public function afterClean(Cleaner $cleaner, array $requestData)
     {
         if (!Search::isNostoSorting($requestData) || !Search::hasCategoryFilter($requestData)) {
-            $this->logger->debugWithSource(
+            $this->debugWithSource(
                 'Nosto sorting not used or not found from request data',
-                $requestData,
-                $this
+                $requestData
             );
             //remove nosto_personalised in case it's a search page
             Search::cleanUpCmpSort($requestData);
@@ -103,21 +109,18 @@ class RequestCleaner
             } elseif ($this->containsGraphQlProductSearchQueries($requestData)) {
                 $this->graphqlHandler->handle($requestData);
             } else {
-                $this->logger->debugWithSource(
+                $this->debugWithSource(
                     sprintf(
                         'Could not find %s from ES request data',
                         self::KEY_BIND_TO_QUERY
-                    ),
-                    $requestData,
-                    $this
+                    )
                 );
                 return $requestData;
             }
         } catch (Exception $e) {
-            $this->logger->debugWithSource(
+            $this->debugWithSource(
                 'Failed to apply CMP - see exception log(s) for details',
-                $requestData,
-                $this
+                $requestData
             );
             $this->logger->exception($e);
         } finally {
