+++ Plugin/Framework/Search/Request/WebHandler.php
@@ -123,13 +123,11 @@ class WebHandler extends AbstractHandler
     public function parseLimit(array $requestData)
     {
         if ($this->pageSize != -1) {
-            $this->getLogger()->debugWithSource(
+            $this->debugWithSource(
                 sprintf(
                     'Using DI value (%s) for the page size',
                     $this->pageSize
-                ),
-                [],
-                $this
+                )
             );
 
             return $this->pageSize;
